+++ RELEASE_NOTES.md
@@ -3,6 +3,9 @@
 ### Bugfixes
 * [#573] Fixed an issue with item attack bonuses not being calculated if added manually via the roll dialog.
 
+### Enhancements
+* [#220] Added ability to drag attacks, spells, abilities, potions, and light sources to the Hotbar.
+
 ## v.1.5.1
 
 ### Bugfixes